+++ packages/homepage/src/pages/react-hooks.js
@@ -1,7 +1,5 @@
 import React from 'react';
 import { graphql } from 'gatsby';
-import TextLoop from 'react-text-loop';
-import styled from 'styled-components';
 
 import TitleAndMetaTags from '../components/TitleAndMetaTags';
 import PageContainer from '../components/PageContainer';
@@ -9,39 +7,9 @@ import { Heading2 } from '../components/headings';
 
 import Layout from '../components/layout';
 import SandboxModal from '../screens/explore/SandboxModal';
-
+import ShuffleWords from './_shuffleWords';
 import WideSandbox from '../screens/explore/WideSandbox';
 import { Container, Sandboxes } from './_explore.elements';
-
-const Words = styled.section`
-  color: white;
-  font-size: 2em;
-  font-family: 'dank mono', 'dm', monospace;
-  text-align: center;
-  margin: 2rem;
-  margin-bottom: 6rem;
-`;
-
-const shuffle = a => {
-  for (let i = a.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [a[i], a[j]] = [a[j], a[i]];
-  }
-  return a;
-};
-
-const words = [
-  'State',
-  'Effect',
-  'Ref',
-  'Context',
-  'Reducer',
-  'Memo',
-  'Callback',
-  'ImperativeHandle',
-  'LayoutEffect',
-  'DebugValue',
-];
 
 export default class extends React.PureComponent {
   state = {
@@ -140,16 +108,7 @@ export default class extends React.PureComponent {
             >
               React Hooks Examples
             </Heading2>
-            <Words>
-              import {'{ '}
-              use
-              <TextLoop interval={1500}>
-                {shuffle(words).map(word => word)}
-              </TextLoop>{' '}
-              {'  }'} from {"'"}
-              react
-              {"'"}
-            </Words>
+            <ShuffleWords />
             <Sandboxes>
               {sandboxes.map(sandbox => (
                 <WideSandbox
